计算收益，防止money字段溢出

+++ logic-server/src/main/java/logic/server/service/impl/action/ChallengeBossSuccessExecutor.java
@@ -44,7 +44,7 @@ public class ChallengeBossSuccessExecutor implements BaseExecutor<ChallengeBossS
         userBossDTO.setBeat(true);
 
         CfgBossDTO cfgBossDTO = CfgManagerSingleton.getInstance().getCfgBossByIdFromCache(arg.getBossId());
-        long moneyIncome = arg.getMultiple() * cfgBossDTO.getRewardMoneyAmount();
+        long moneyIncome = ((long)arg.getMultiple()) * ((long)cfgBossDTO.getRewardMoneyAmount());
         long finalMoney = userDTO.getMoney() + moneyIncome;
         userDTO.setMoney(finalMoney);
